Feature/refactor (#4)
* Limpieza código
* Refactorizacion TemplateService
* Refactor inyección de configuraciones
