+++ Core/Interfaces/TemplateService.cs
@@ -1,0 +1,8 @@
+using sds.notificaciones.core.entities;
+using sds.notificaciones.core.DTO;
+
+namespace sds.notificaciones.core.Interfaces {
+    public interface TemplateService {
+        Mail GenerateMail(Notificacion notificacion);
+    }
+}
